address comments
- [x] Renamed files and types to drop mention of rest as rest is now the only implemented api request type.
- [x] Add validation for input `type` parameter
- [x] implement instance of octokit in the factories and pass to instances of strategies through construction parameter - this also removes inheritance on the BaseStrategy type.
- [x] Renamed BaseStrategy to PackagesClient
- [x] Added extra unit test for PackageType parameter
- [x] Combined both process files into a single file.
- [x] Renamed `processRestResponse` to `processResponse`
- [x] Removed dist folder from commit.
