+++ src/delete/delete.strategy.factory.ts
@@ -1,7 +1,9 @@
 import { DeleteStrategy, Input } from "../types"
-import OrganizationRestDeleteStrategy from "./rest/organization.rest.delete.strategy"
-import UserRestDeleteStrategy from "./rest/user.rest.delete.strategy"
+import OrganizationDeleteStrategy from "./strategies/organization.delete.strategy"
+import UserDeleteStrategy from "./strategies/user.delete.strategy"
+import {PackagesClient} from "../clients/packages.client";
 
 export function decideDeleteStrategy(input: Input): DeleteStrategy {
-  return input.organization ? new OrganizationRestDeleteStrategy() : new UserRestDeleteStrategy()
+  const packagesClient = new PackagesClient(input);
+  return input.organization ? new OrganizationDeleteStrategy(packagesClient.client) : new UserDeleteStrategy(packagesClient.client)
 }
