+++ src/query/query.strategy.factory.ts
@@ -1,7 +1,9 @@
 import { Input, QueryStrategy } from "../types"
-import OrganizationRestQueryStrategy from "./rest/organization.rest.query.strategy"
-import UserRestQueryStrategy from "./rest/user.rest.query.strategy"
+import OrganizationQueryStrategy from "./strategies/organization.query.strategy"
+import UserQueryStrategy from "./strategies/user.query.strategy"
+import {PackagesClient} from "../clients/packages.client";
 
 export function decideQueryStrategy(input: Input): QueryStrategy {
-  return input.organization ? new OrganizationRestQueryStrategy() : new UserRestQueryStrategy()
+  const packagesClient = new PackagesClient(input);
+  return input.organization ? new OrganizationQueryStrategy(packagesClient.client) : new UserQueryStrategy(packagesClient.client)
 }
