+++ test/delete/delete.strategy.factory.spec.ts
@@ -1,6 +1,6 @@
 import { decideDeleteStrategy } from "../../src/delete/delete.strategy.factory"
-import OrganizationRestDeleteStrategy from "../../src/delete/rest/organization.rest.delete.strategy"
-import UserRestDeleteStrategy from "../../src/delete/rest/user.rest.delete.strategy"
+import OrganizationDeleteStrategy from "../../src/delete/strategies/organization.delete.strategy"
+import UserDeleteStrategy from "../../src/delete/strategies/user.delete.strategy"
 import { Input, PackageType } from "../../src/types"
 
 test("decide default delete strategy", () => {
@@ -8,14 +8,15 @@ test("decide default delete strategy", () => {
     names: ["test"],
     versionPattern: /.*/,
     keep: 0,
-    token: "",
+    token: "abc-123",
     dryRun: false,
     user: "",
     organization: "",
     owner: "SmartsquareGmbH",
+    type: PackageType.Npm,
   }
 
-  expect(decideDeleteStrategy(input)).toBeInstanceOf(UserRestDeleteStrategy)
+  expect(decideDeleteStrategy(input)).toBeInstanceOf(UserDeleteStrategy)
 })
 
 test("decide rest user delete strategy", () => {
@@ -24,14 +25,14 @@ test("decide rest user delete strategy", () => {
     versionPattern: /.*/,
     keep: 0,
     type: PackageType.Npm,
-    token: "",
+    token: "abc-123",
     dryRun: false,
     user: "user",
     organization: "",
     owner: "SmartsquareGmbH",
   }
 
-  expect(decideDeleteStrategy(input)).toBeInstanceOf(UserRestDeleteStrategy)
+  expect(decideDeleteStrategy(input)).toBeInstanceOf(UserDeleteStrategy)
 })
 
 test("decide rest organization delete strategy", () => {
@@ -40,12 +41,12 @@ test("decide rest organization delete strategy", () => {
     versionPattern: /.*/,
     keep: 0,
     type: PackageType.Npm,
-    token: "",
+    token: "abc-123",
     dryRun: false,
     user: "",
     organization: "org",
     owner: "SmartsquareGmbH",
   }
 
-  expect(decideDeleteStrategy(input)).toBeInstanceOf(OrganizationRestDeleteStrategy)
+  expect(decideDeleteStrategy(input)).toBeInstanceOf(OrganizationDeleteStrategy)
 })
